refactor: user domain 내의 type 폴더 추가
