--- src/main/java/org/colcum/admin/domain/user/domain/Branch.java
@@ -1,5 +1,0 @@
-package org.colcum.admin.domain.user.domain;
-
-public enum Branch {
-    JONGRO
-}
