+++ src/main/java/org/colcum/admin/domain/user/domain/type/Branch.java
@@ -1,0 +1,5 @@
+package org.colcum.admin.domain.user.domain.type;
+
+public enum Branch {
+    JONGRO
+}
